Replaces source with rendered code in docs
According to this article: https://storybook.js.org/docs/react/writing-docs/doc-blocks#docspage-1

There is a way to configure stories to display generated snippets on the DocsPage instead of source. Maybe we can find a way to apply this automatically to all stories; in a template perhaps? Something to work off of.