Changes style logic for link icon
With this change, the (external link) icon inside the link component
will size itself based on its accompanying text size, with a minimum to
ensure it does not fall below 20px.

This also removes some confusion in what the correct values of size are
expected to be. Before, one may assume `xs` was the expected value
(valid for icon but not for text).