feat(icon): renames Help to HelpOutline
To keep with convention for the rest of the icon library, `help` is no
longer a valid value for the name prop- it has been replaced with
`helpOutline` to go along with the introduction of `helpFilled` to the
library.

BREAKING CHANGE: name="help" is no longer valid; use name="helpOutline"
instead.